pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = ""
authors = [
    {name = "Kevin Leung",email = "keleung@microsoft.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "flake8 (>=7.3.0,<8.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "azure-identity (>=1.25.0,<2.0.0)",
    "azure-storage-blob (>=12.19.0,<13.0.0)",
    "aiohttp (>=3.12.15,<4.0.0)",
    "semantic-kernel (>=1.12.0,<1.15.0)",
    "pydantic (>=2.5.0,<2.9.0)",
    "python-docx (>=1.1.0,<2.0.0)",
    "docxtpl (>=0.18.0,<1.0.0)",
    "pyqt5 (>=5.15.11,<6.0.0)",
    "markdown (>=3.9,<4.0)",
]

packages = [
    { include = "agents", from = "src" }
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
